fix(markdown-native): isolate resolver and honor app ABIs
