+++ crates/supramark-markdown/packages/react-native/android/build.gradle
@@ -29,6 +29,14 @@ def safeExtGet(prop, fallback) {
     rootProject.ext.has(prop) ? rootProject.ext.get(prop) : fallback
 }
 
+// Honor the app-level `reactNativeArchitectures` Gradle property so a
+// single-ABI app configures only the native library it needs. The fallback
+// remains all four ABIs staged by scripts/prepare-native.js.
+def nativeArchitectures(defaults) {
+    def value = rootProject.findProperty("reactNativeArchitectures")
+    return value ? value.split(",").collect { it.trim() }.findAll { it } : defaults
+}
+
 apply plugin: "com.android.library"
 apply plugin: "com.facebook.react"
 
@@ -47,12 +55,8 @@ android {
             }
         }
 
-        // Keep in sync with scripts/build-native.sh + prepare-native.js,
-        // which build and stage all four ABIs into jniLibs/. CMake
-        // FATAL_ERRORs if a filtered-in ABI has no staged .so, so this
-        // list must not exceed what was staged.
         ndk {
-            abiFilters "arm64-v8a", "armeabi-v7a", "x86_64", "x86"
+            abiFilters.addAll(nativeArchitectures(["arm64-v8a", "armeabi-v7a", "x86_64", "x86"]))
         }
     }
